Format

+++ priv/repo/migrations/20240405212253_create_clientes_table.exs
@@ -4,7 +4,7 @@ defmodule RinhaBank.Repo.Migrations.CreateClientesTable do
   def change do
     create table(:clientes) do
       add :nome, :string, null: false
-      add :limite, :integer, null: false
+      add :limite, :bigint, null: false
       add :saldo, :bigint, null: false
     end
   end
